Big changes in nsParameters, nsGlobal and controllers.
nsGlobalParameters is a new class that contains the global.nml as defaults and the experiment file as the only condition. ALl other objects ask nsGlobal for information and do not create new links to the nml file. 

Controllers need to query the relevant properties of the nml file in their beginSession() functions. 

nsNMLFile is now a simple class to open an nml file ,check its format and possibly include another file and merge it.