Created the prepareExperiment function to allow users to interact with the experiment before it starts running.
Added pause button and functionality.

Changed stage to a local variable in the nsControl::idle loop and allow other program parts to use nsFlags only

FIxed a bug in the asChar()  function of nsVariables.cpp

Changed the deinfition of nsParameters::getString to match the order of parameters as getDouble etc.

Renamed nsGraphics to nsEyelinkSetup