Fixed some bugs to allow default parameters in a getParm call to use a const char*.
ADdded functionality to switch off EPS logging in AO plugin

Fixed a bug in behavior that tried to draw fixation windows for stimuli that no longer existed at the end of a trial.

Fixed a bug that started a stimulus with frame>1 if the stimulus was supposed to start immediately with fixation onset. 

Added code to check for frame skew that would lead to inaccurate timing

Changed key presses for controller to always require ctrl. (ctrl-space, ctrl-return).