Removed Eyelink from the GUI. Users can use keys to control Eyelink, or the eyelink gui itself.
Fixed a bug that kept Eyelink calibration running when neurostim tried to exit.

moved Key stroke handling of GUI, & Windows to the same place in the controller. This improves consistency... 

Keyboard listeners now also can ask the 'alt' and 'shift' state (boolean). Ctrl- presses are not sent to KBListeners (but to the controller).