Changes to improve the remoteCommand interface.
Main change is that return messages are all wrapped in an XML tag equal to the command that generated it (for consistency checking on the other side) and that all commands are terminated by a \r\n carriage-return line feed. This simplifies the parsing of messages on teh receiving side. Tested with nsOnline in matlab and messages are coming across ok. 

Not yet tested whether old remoteCommands for stimuli are still being passed properly.