Finally figured out how to make stimcon parse factorials recursively. This now allows all levels of nesting for factorial definition.
Parameters with _ or ( or  ) in them are type=Function by default. 

Added functionality to allow nsBackground to set a subset of nsGlobal parameters. This implies that this can be done per condition and in factorial definitions. 
Parameters include TRIALDURATION, BACKGROUND RGBA, REWARD etc. 

Added a beginTrial function to nsGui to allow updates from nsBackground. 

No recompilation should be necessary.