Removed 'visible' - can't easily implement it across stimuli without adding it to all, which makes it not user-friendly.
Added rotation, scaling and transformation per stimulus, so X and Y positions are adjusted using openGL instead of within the stimulus itself. Changed all stimuli to be compatible with this. Combined the setup openGL commands into a function in cic.

Added a mouse stimulus which tracks the position of the mouse (to be logged at the end) and tracks all mouse clicks.