Added firstFrame event to cic - notifies immediately after first frame.
Edited output so that relevant data is saved; added trial start and end times experiment end time in cic for data collection.

Changed myConfig to a function.

Edited the plugin file so listeners can be removed; used this in the mouse stimulus so a single matrix of trajectory values can be logged after every trial instead of constant updating.