Changed cic so that all draw functions draw into an offscreen window, which is then drawn to the onscreen window between beforeFrame and afterFrame events.
GUI lists keys, a text feed which can be accessed by cic.gui.writeToFeed('text'), a list of pre-set parameters and a copy of the experimental window. Added a property in cic which flags if the gui is on; it writes to gui when frames are dropped. This is currently causing timing problems/frame drops.