Some changes.
Changed RSVP in stimulus to recalculate on/off times instead of preallocating. Additionally, this reshuffles whenever it reaches the end of the condition list which was input.

Added a new function to plugins which can be used to add new keys which evaluate to specific functions. Currently, this will be overridden when the keyboard() function is used within the plugin itself.

Removed luminance as a property, color is now a length 3 vector.