Key handling changes.
Adjusted keys so they can be added using the o.addKey(key,@fnHandle) format. This is a function in the plugin class which calls the associated function when a key is pressed. This replaces the previous keyboard handler and has been updated in relevant subclasses.

Some other minor functionality fixes.