Fixed bugs in factorial - nrLevels /conditons were not evaluated properly for stimuli with multiple changing parameters.
Added functionality to allow users to specify parameter values as vectors, not as cell arrays i.e. 1:3 will be interpreted as {1,2,3} , which makes writing factorials a bit easier.
Added functionality to allow some parameters to have a single value in a factor even if the number of levels is larger than 1 (the single level of that property is then used for all levels, similar to basic Matlab scalar/vector functionality).
Wrote a first analysis script (anaTae) that uses the output structs.