Improved handling of dual screens
Changed how RSVP's are handled internally and how they are specified.
The new version has less bookkeeping overhead. The only thing that does not work yet is to have differnt RSVPs for different
conditiosn (i.e. rsvp is not  a propoerty). That will be next.
Added sog as an RSVP example