Substantial change to the way blocks/factorials and conditions are used and nested.
Previously the information was stored in facotrials and then extracted (and duplicated) by CIC before the start of the experiment. This seems inefficient. Now the conditions specifications stay in the factorials and the blocks maintain a counter of the current trial and index into the list of the conditions in the relevant factorials.
All changes are behind the scenes.