Removed the 'duration' property from the feedback items in the feedback class.
The idea of a 'duration' doesn't seem to apply to all modes of feedback. At present
the only feedback class that uses the duration property of the feedback items is
the liquid class. I've modified the feedback and liquid classes so that the duration
property is handled locally by the liquid class.