Checked the re-assigning values to function properties:
f.X = '@f.Y +1'
f.X = '@f.Z +1'
now works. Detecting that this is a request for a reassignment uses dbstack which is a bit of a kludge, but it seems to work ok and not be too slow.

Updated some of the code comments and updated the duplicate function to take this change into account.

Reinstated duplicate example in tae.