Introduced a parameter class that handles the @functions , logging, and setting/getting parameter values.
This cleans up some of the plugin code, and stores a log with each parameter (which hopefully will make retrieval easier too; not implemented yet).
This is functional, the only changes in the rest of the code are the 'SetAccess' restrictions we had on properties before; they are now gone (have to be public).

Removed postprocessing in parameters - was not used.