Removed the obligatory output plugin being added in cic and replaced it with saving of the cic object. This change would break any existing experiments that use the old plugin format for output. If they exist, they can either add the output plugin in their script or switch to Bart's new get() format.
Currently very rudimentary and probably not very clever. Saves on exit, and also on deconstruction. We've lost the "save every n trials" functionality that existed in the plugin version. Also, c.file, c.fullFile, and c.fullPath (new) exist, but it feels like these should be within the c.dirs structure (or that this is instead renamed c.output).