Revamped keyboard handling to allow two keyboards and explicit keyboard specification (which appears to be ncessary on Linux/Ubuntu) .
Plugins now call addKey() to add a key and define a keyboard(plg,key) function to handle keys for that plugin.

Stimulus keys can be handled by a separate keyboard (cic.kbInfo.subject) than the keys for plugins (cic.kbInfo.experimenter)

If these are not set, the default keyboard (which appears to be both keyboards in Windows, but only one in Linux) will handle the keys.

Users can set kbInfo.subject and kbInfo.experimenter

kbInfo.pressAnyKey is the keyboard that allows to start the experiment (and block). It can be set to -1 (any) or a specific keyboard.