Moved vsync and other advanced parametres into a .timing struct so that they can be set at runtime (e.g. in myRig)
Removed deadline from the call to flip.
Vsyncmode ==0 is a busy wait for a flip. So if the vbltime is correct, then a drop has occurred only if the vbltime is more than 1 frame late. Hardcoded this with 10% slack.
For vsyncmode==1 or 2, the user can specify the frameslack as a percentage (as before)