Changed default plugin order to run plugins before stimuli (that way an eyelink calibration can start before anything else happens in teh first trial, for instance).
This order seems to make sense as plugins are often peripherals that need setup first. Of course a user can change the order manually.

Documented framedrop math.

Moved the AFTERFRAME code to the end of the frame loop (after the FLIP) .  This appears to improve performance substantially.

Documented performance issue in starstim
Fixed file save location issue in starstim