Restricted simulated nafcResponse to times when the real observer woul dbe allowed to answer.
Added latin squares designs to randomization of blocks.
nrConidtions  now refers to conditions in this block (summing over blocks makes no sense as some of those conditions could be the same).
Added an alwaysOn property to a stimulus. Together with cic.itiClear this allows stimuli to stay on even when their ontime is relative to fixation onset.
In the current version the .startTime is set only once, not every trial ,for an .alwaysOn stimulus that makes sense to me (becuase it starts only at the beginning of the experiment), but could make analysis a bit trickier.