Added operator functionality to adaptive parameters so that v  = jitter1+jitter2  returns the sum of the current values of the jitter object.
This allows users to set a property to an adaptive parameter and use those properties in a neurostim or matlab function. |This kind of adaptive parameter is updated after every trial so it
makes sense only when the adaptive parameter is not specific to a condition/design.  I think that this mainly helps using jitter in an easier way. Becuase the value of the adaptive object is returned only when an operator is used,
s.X returns the adaptive object while +s.X  (=uplus(s.X)) returns the value of the adaptive object.

When a property that has an adaptive parameter as its value is logged, I pull out the value (instead of logging the entire adaptive parameter).

Updated adaptiveDemo to this new usage.