Working glshaderDemo.m.
The fragment shader doesn't need to know the number of randels. It
just takes the value from the red channel of Image and uses it to
interpolate into clut.

It is up to the external code (i.e., glshadertest in this case) to
make sure the values written to Image are in the range 0 to
nRandels, and that clut is atleast that many elements wide.

This seems to work well on my MacBook (which supports full 32 bit
float textures) for a wide range of nRandels.