More fixes for parameter. My previous fix back-dated the trial time but not the clock time. If we are going to do one, we must do both.
Also ensured that in these cases (where the logged data is actually a time) there is no ambiguity by returning NaNs for the "data" output, forcing use of trialTime or time. That is, this is an event rather than a logged param change.

None of this is particularly elegant. Perhaps neurostim.parameter should have an "event" sub-type for these special cases?