Reverted changes to the "experiment time" clock that were introduced a while back.
Changing that clock on a per-trial /event basis means that time is not running on a fixed increment, whcih creates havoc when syncing data with another device.

I think all that was needed in the parameter.get()  was to backdate the times at which startTime and stopTim occur for stimuli from their time to the data that was stored with the event.
That improves timing of those events as it synced to the screen flip.

I removed the code that set the data fields to NaN - those raw data can come in handy for trouble shooting.
And the adjustment of firstFrame is now superfluous; its trialTime  is always zero.

With these changes, I can sync properly with an external device (ripple), without it that was not possible (apparent clock drift  of 100's of ms over the course of an experiment).
Of course these changes may break some other dependent analsysi code.. this should be tested before this branch can be merged back into the master.