Fixed clearing of the overlay.
The problem here was that the plugin class calls glTranslate to set
the draw origin to the centre of the screen, so the overlay would
clear properly only up until the first call to beforFrame(), after
which it would only clear one quadrant of the screen.

One fix, implemented here, is to push the current (default?) matrix
before calling beforFrame(), and then pop it afterward.

Calls to clearOverlay() (or FillRect) in the frame loop then behave
as expected.