text stimulus does not seem to be working properly.
I only needed text feedback so i changed that to use a call to cic.drawFormattedText in textFeedback instead  - simpler, I think.
The only loss of functionality is specifying X and Y in cm on the screen for the text. This shoud now
be done using the left and top arguments, and the winRect as in PTB's DrawFormattedText.