Replaced the feedCache in cic with a dedicated logger object. This object does what the feedcache did (i.e. write to the command line using
color if requested) , but it can also send those same messages to a remote machine running matlab and show them on the command line there (using TCP).

The advantage is that this allows using a single monitor for the experiment and still get info to the experimenter.
Currently the remote is passive (it only shows the feed). It could be extended with some control as well (i.e. start /stop  the experiment).

The only change in the other code is that plugin.writeToFeed now wants a string and can no longer to sprintf style formatting on the fly.
The devloper can easily do this in the plugin, using sprintf so no real loss there...I updated existing usage of writeToFeed.