Fixed a bug (?) in str2fun() due to partial matches of property names.
Consider a plugin with two properties: plg.prop, and plg.prop1. Attempting
to use plg.prop and plg.prop1 in a function string, something like:

  '@plg.prop + plg.prop1'

should be expanded as:

  '@args{1}.getValue() + args{2}.getValue()'

but is/was expanded as:

  '@args{1}.getValue() + args{1}.getValue()1'

due to regexprep replacing all instances of the first reference plg.prop,
including the partial match with plg.prop1.

The solution implemented here is to replace the references explicitly,
one at a time, left to right.