Big change in CIC/run:
Added AsyncFlip to squeeze out a bit more time for stimulus processing that is otherwise used to wait for
the VBL.
This version works without drops for many experiments.

The downside is that the afterFrame function is now processed at an undefined moment in time (flip may or may not have happened).
And, more importantly, when afterFrame runs, stimulus startTimes will not have been updated yet for the currently flipping frame.
This could be fixed in user code (by moving it to beforeFrame..)

Checked that PsyncTimeoutFrames,1 is needed for VPIXX.