Cast output from extractAfter() to char() in plugin.localizeParms().
Apparently, as of R2017a, the output type of extractAfter() and
friends depends on the type of the first input argument, but in
R2016b extractAfter() always returned string arrays. This caused
the dynamic structure reference to fail when localizing plugin
parameters.

This fix explicitly casts the output from extractAfter() to
char(). This should have no effect in Matlab versions after R2017a
but makes it possible to run on R2016b as well.

Note that extractAfter(), startsWith() etc. were introduced in
R2016b, so this code will fail on versions prior to R2016b.