noiseclut.m now supports a separate sampling distribution/function (e.g. Gaussian noise) for each luminance/color channel (e.g. lum only, RGB, LA, RGBA, etc.). For example, can be used to make sparse Gaussian luminance noise (see SPARSE demo in noiseGridDemo.m).
This commit adds an alpha channel in the CLUT. This required updating the shader. The final alpha of a pixel is now the product of its value in the spatial alpha mask in the color-by-numbers image, and the alpha channel of the CLUT entry it points to.
Updated demos. Some changes to names of variables and parameter specification in the experiment file.