Changed calling of flipCallbacks in VsyncMode==0 and Vsyncmode==1
In this new version:

in BOTH vsyncModes:
 flip times (.startTime ) always reflect the actual time of the flip.
 stimulus.onsetFunction and stimulus.offsetFunction are always called AFTER the flip has occurred,

in the SYNC mode (0)
flip callbacks happen BEFORE the afterFrame function. In theory that means
an afterFrame function could use .startTime (although this is not necessary or recommended).
onset/offsetFunctions are closer in time to the actual flip time.

in ASYNC mode (1)
flip callback happen AFTER the afterFrame function.  So afterFrame should not use .startTime or .stopTime
onset/offsetFunctions are a little bit later compared to the actual flip time than in SYNC mode. (because afterFrame is called first).

With this change the SYNC mode behaves the same way as older versions and ASYNC mode is new, intended to get a little bit of performance improvement by blocking only at the last possible moment in the frame loop.