Fix for the LINLUT color mode.
The LINLUT color mode is broken (at last on Linux, and I believe also on Window) because
ScreenDacBits() returns 0 (always). This is likely related to:

https://groups.yahoo.com/neo/groups/psychtoolbox/conversations/messages/22838

This fix reads the current gamma table to determine the number of slots in the hardare lookup table and
computes the inverse gamma table to suit.