Initial version of a new way to implement a flash that can be recorded with a diode
to confirm timing.
The main problem with the old approach was that the diode background and the persistence of that background during the ITI
could not be controlled separately from the stimulus. stimulus

While doing this I also fixed a bug with itiClear =false in vsyncMode==1 ; those ITIs were cleared regardless. Fixed now.