Made neurostimEyelinkDispatchCallback the default callback for all Eyelink plugin. This now handles eyelink sounds with PsychPortAudio, as well
as the drawing of targets to an overlay when a VPIXX or similar is used.  This fixes hardware control battles between sound (Snd) and PsychPortAudio (sound plugin).

Added a beep funciton to the sound plugin to replace the Beeper used by the original Eyelink Dispatch callback.
sound.beep is possibly not optimal in terms of timing, but good enough for eyelink sound feedback.