Add a warning when loading a cic object from a crashed experiment.
If neurostim crashes (or otherwise exist unexpectedly), the file
you're left with is a partial record of the experiment. It is possible
that additional trials were run that you now have no record of. This
can cause confusion when trying to reconcile data or events that may
have been recorded elsewhere (eye tracker, daq system etc.). This
change adds code to issue a warning when loading a partial save to
minimise confusion.