Moved the code to update a saved object to the new (current) classdef to a shared function in plugin.
This requires giving the plugin class access to some of the (previously) protected properties of its child classes.
What is really new  (and not jsut moved) in this code is the re-establishing of dynamic properties at load time.
Probably because objects that no longer match the current classdef are converted to structs, their dynprops are lost.
I now recreate those, and make them read only.  This avoids errors in plugin code that relies on the dynprops.