default color of gabor
The default color inherited its value from stimulus.m This could produce a non-sense gabor patch if there was a mismatch of color from stimulus.m and colormode from cic.
Hence in this commit, the default color is defined within gabor.m
Also a note on the meaning of color is added.