Create cic.trialSuccess that evaluates to the sucess fo the current trial based on all or a subset of required behaviors.
Used in block.afterTrial  and adaptive.afterTrial

Added an option to only update an adaptive parameter when the success of a set of specified behaviors evaluates to true.
This is necessary to ignore updates to adaptives in trials in which the behavior is incomplete (e.g. a to ignore a key press after a fixation break)