Fix updating of adaptive plugins re: commit 6b6e581.
Commit 6b6e581 modified the adaptive base class to
call it's update() method in beforeTrial rather than
after trial. However, beforeTrial is only called on
plugins *after* beforeTrial is called on the current
block/design object. As a result, the design object
cannot know the new/updated value that should be
assigned for the next trial.

This change calls beforeTrial on any adaptive plugins
before calling beforeTrial on the current block/design.