Fixed a bug in jittered parameter updating.
Previously a design with conditions only (i.e. no fac1) would not update jitter objects and could remove adaptive parameters specified in the design.
The design object now handles this situation correctly and also allows multiple conditiions to be assigned the same value.

It is still somewhat odd that
	gabor.X = jitter

design.conditions(1).gabor.orientation = 45
design.conditions(2).gabor.X = 0

will ignore the jittered X not only in condition 2, but also in condition 1 (as soon as the specs have some value for gabor.X, the jitter is ignored).