Added two new plugins.
mdaq uses the Matlab Data Acquisition toolbox to interact with daq devices (e.g., National Instruments); it is called mdaq to distinguish it from the daq plugin which uses PTB tools to talk to the hardware.
scanbox  is a plugin that interacts with the NeuroLabware two-photon imaging hardware running the Scanbox software.
The circularBuffer util is used by mdaq to show recently acquire data in scope like fashion on the nsGui.

Early beta versions.